Latest hyperHTML with proper Custom Elements initialization
Due to issue [#671](https://github.com/w3c/webcomponents/issues/671)
it is not possible to synchronously set attributes passing
through the custom element prototype when an element
comes from a template and it's not already live.

Previous version of hyperHTML was somehow cheating and
attaching properties to the element directly,
shadowing its inherited behaviour after upgrade to CE.

The issue is explained in here:
https://github.com/w3c/webcomponents/issues/671#issuecomment-332867127

This PR uses live nodes instad of template one
so that the test passes without needing any special tratment.