Revise heading structure to improve accessibility
Previously the headings were repetitive and did not list the frameworks.
This could make it difficult to navigate with a screen reader.
It could also make it harder for search engines to understand the content.

To improve this a couple changes were made:
- Framework names were promoted to `h3`s
- "No issues found" is no longer a heading

Luckily all heading styling was applied via classes so no CSS changes were needed.