Added goldens for expected summary results
Added a meta/expectedResults.json file for each library that includes the summary field of results.json, and adds an assertion as part of index.js that the expected summary matches the actual summary.

This way, no change can accidentally change the results without us seeing, and any change that intends to change results expresses that in a way that's easy to see in code review.

This was prompted by stencil's build failing in such a way that it still generated results, but those results were just a failing 0% score because of a build configuration problem.